Add city route ordering with route_order; NearbyCitySelector, map & afisha UI updates

+++ database/migrations/2026_05_07_000000_add_route_order_to_cities_table.php
@@ -9,7 +9,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('cities', function (Blueprint $table) {
-            $table->unsignedSmallInteger('route_order')->nullable()->after('population');
+            $table->unsignedInteger('route_order')->nullable()->after('population');
         });
     }
